FEAT: corrected code to requested changes
